chore: migrate from runTest to runInlineTest
This is a preparation for collecting fixtures from the file system,
where we cannot allow various assets to coexist in the same dir.
