--- test/assets/export-2.fixtures.ts
@@ -1,29 +1,0 @@
-/**
- * Copyright (c) Microsoft Corporation.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-import { folio as baseFolio } from '../..';
-
-const fixtures = baseFolio.extend<{ testTypeOnly: string }, { workerTypeOnly: number }>();
-
-fixtures.testTypeOnly.init(async ({}, runTest) => {
-  await runTest('testTypeOnly');
-});
-
-fixtures.workerTypeOnly.init(async ({}, runTest) => {
-  await runTest(42);
-}, { scope: 'worker' });
-
-export const folio = fixtures.build();
