--- test/assets/test-error-visible-in-fixture.ts
@@ -1,28 +1,0 @@
-/**
- * Copyright (c) Microsoft Corporation.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-import { folio } from '../../';
-
-const builder = folio.extend<{ postProcess: string }>();
-builder.postProcess.init(async ({testInfo}, runTest) => {
-  await runTest('');
-  console.log('ERROR[[[' + JSON.stringify(testInfo.error, undefined, 2) + ']]]');
-});
-const { it, expect } = builder.build();
-
-it('ensure fixture handles test error', async ({ postProcess }) => {
-  expect(true).toBe(false);
-});
