--- test/assets/typescript.ts
@@ -1,27 +1,0 @@
-/**
- * Copyright (c) Microsoft Corporation.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-import { folio } from '../..';
-const { it, expect } = folio;
-import './global-foo';
-
-it('should find global foo', () => {
-  expect(global['foo']).toBe(true);
-});
-
-it('should work with type annotations', () => {
-  const x: number = 5;
-  expect(x).toBe(5);
-});
